% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascii}
\alias{ascii}
\title{ascii}
\format{
data.frame size 94x4
}
\usage{
ascii
}
\description{
Equivalences of ascii code. Columns are Number, (ascii) Symbol, Q (score) and P (probability of error)
}
\keyword{datasets}
