% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascii_v}
\alias{ascii_v}
\title{ASCII code}
\format{
Named character vector of length 94. Names are ascii character and value is the probability of error.
}
\usage{
ascii_v
}
\description{
Vector ascii
}
\keyword{datasets}
