% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prior_errors.r
\name{calculate_prior_errors}
\alias{calculate_prior_errors}
\title{Computes prior probability of errors}
\usage{
calculate_prior_errors(
  counts_pnq,
  output_file = NULL,
  reference_sequence,
  save = FALSE
)
}
\arguments{
\item{counts_pnq}{Data frame with columns position nucleoide quality counts, as returned by parse_nucleotides_per_qscore.}

\item{output_file}{File name for output, if save=TRUE.}

\item{reference_sequence}{Reference sequence: vector of characters, as returned by load_reference_sequence}

\item{save}{Logical. Should data be saved in a output_file?}
}
\value{
Data frame with columns position nucleotide prior.error.
}
\description{
This is an auxiliary function in single package. It takes a data frame with counts by position, nucleotide and Qscore and it summarises it into proportion of nucleotide counts by position.
}
\examples{
ref_seq_file = system.file("extdata", "ref_seq.fasta", package = "single")
ref_seq = load_reference_sequence(ref_seq_file)
counts_pnq_file = system.file("extdata", "example_train_countsPNQ_parsed.txt", package = "single")
counts_pnq = read.table(file=counts_pnq_file, header=TRUE)
calculate_prior_errors(counts_pnq,reference_sequence=ref_seq)
}
