% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_consensus_sequence.r
\name{compute_consensus_sequence}
\alias{compute_consensus_sequence}
\title{Compute consensus sequence}
\usage{
compute_consensus_sequence(df, cutoff_prob = 0.2)
}
\arguments{
\item{df}{data.frame with the columns: nucleotide, probability, position}

\item{cutoff_prob}{Numeric. Nucleotides with probability below this number will be removed from consensus computation.}
}
\value{
Character vector, consensus sequence
}
\description{
This is an auxiliary function in single package. It computes consensus from a data.frame as the one returned by single_evaluate_model()
}
