% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_seq.r
\name{consensus_seq}
\alias{consensus_seq}
\title{Compute consensus sequence}
\usage{
consensus_seq(df, cutoff_prob = 0.2)
}
\arguments{
\item{df}{data.frame with the columns: nucleotide, probability, position}

\item{cutoff_prob}{Numeric. Nucleotides with probability below this number will be removed from consensus computation.}
}
\value{
Character vector, consensus sequence
}
\description{
This is an auxiliary function in single package. It computes consensus from a data.frame as the one returned by single_evaluate()
}
\examples{
require(dplyr)
seqs = system.file("extdata", "single_reads_ex_corrected.txt", package="single")
seqs_table= read.table(seqs, header=TRUE,nrows=100*10+1)
seqs_table = seqs_table \%>\%
    select(nucleotide,p_right_priors_model,position) \%>\%
    rename(probability=p_right_priors_model)
consensus_seq(seqs_table)
}
