% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mutations.r
\name{detect_mutations}
\alias{detect_mutations}
\title{Detect mutations beween sequence and reference}
\usage{
detect_mutations(sequence, reference)
}
\arguments{
\item{sequence}{Vector of characters, sequence you want to compare to reference}

\item{reference}{Vector of characters, reference sequence.}
}
\value{
Character vector, in each site the difference between both vectors in the form reference - index - sequence
}
\description{
This is an auxiliary function in single package. It Compares the vectors sequence and reference, and returns the differences.
}
\examples{
detect_mutations(c("A","A","A","A"),c("B","A","B","A"))
}
