% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logregr.r
\name{fit_logregr}
\alias{fit_logregr}
\title{Fit SINGLE's logistic regression}
\usage{
fit_logregr(
  counts_pnq,
  ref_seq,
  prior_error,
  prior_mutation,
  save = FALSE,
  output_file_fits,
  output_file_data,
  verbose = FALSE
)
}
\arguments{
\item{counts_pnq}{Data frame with columns position nucleoide quality counts, as returned by parse_countspnq}

\item{ref_seq}{Reference sequence: vector of characters, as returned by load_ref_seq}

\item{prior_error}{Data frame with columns position nucleotide prior.error, as the one returned by prior_errors().}

\item{prior_mutation}{Data frame with columns wt.base, nucleotide and p_mutation (probaility of mutation), as the one returned by prior_mutations().}

\item{save}{Logical. Should data be saved in a output_file?}

\item{output_file_fits}{File into which save the single fits if save=TRUE}

\item{output_file_data}{File into which save the fitted data if save=TRUE}

\item{verbose}{Logical.}
}
\value{
data.frame with columns position, nucleotide, slope and intercept (of the sigmoidal regression).
}
\description{
This is an auxiliary function in single package. It takes counts_pnq and for each position and nucleotide it fits SINGLE's logistic regression.
}
