% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.predict..r
\name{glm.predict.}
\alias{glm.predict.}
\title{Computes prior probability of mutations}
\usage{
glm.predict.(x, slope, intercept)
}
\arguments{
\item{x}{Numeric. Values to evaluate.}

\item{slope}{Slope of the sigmoidal function to evaluate.}

\item{intercept}{Intercept of the sigmoidal function to evaluate.}
}
\value{
Numeric.
}
\description{
This is an auxiliary function in single package. It evaluates the sigmoidal function given by the parameters slope and intercept on x.
}
\examples{
x = c(-10:10)
y = glm.predict.(x,1,2)
plot(x,y)
}
