% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mismatches.R
\name{list_mismatches}
\alias{list_mismatches}
\title{Lists mismatches between two DNAstrings}
\usage{
list_mismatches(ref, seq)
}
\arguments{
\item{ref}{DNAString, reference sequence.}

\item{seq}{DNAString, target sequence, same length as ref.}
}
\value{
Character vector containing Nucleotide in ref Position Nucleotide in seq. If ref and seq are equal, it returns NA.
}
\description{
This is an auxiliary function in single package, to list the mutations of two DNAstrings.
}
\examples{
ref = Biostrings::DNAString("AAAA")
seq = Biostrings::DNAString("AGAT")
list_mismatches(ref,seq)
list_mismatches(ref,ref)
}
