% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ref_seq.r
\name{load_ref_seq}
\alias{load_ref_seq}
\title{Load reference sequence from fasta file}
\usage{
load_ref_seq(file)
}
\arguments{
\item{file}{fasta file with reference sequence}
}
\value{
The reference sequence as a vector of characters
}
\description{
Load reference sequence from fasta file
}
\examples{
load_ref_seq(system.file("extdata", "ref_seq.fasta", package = "single"))
}
