% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference_sequence.r
\name{load_reference_sequence}
\alias{load_reference_sequence}
\title{Load reference sequence from fasta file}
\usage{
load_reference_sequence(file)
}
\arguments{
\item{file}{fasta file with reference sequence}
}
\value{
The reference sequence as a vector of characters
}
\description{
Load reference sequence from fasta file
}
\examples{
load_reference_sequence(system.file("extdata", "ref_seq.fasta", package = "single"))
}
