% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation_rate}
\alias{mutation_rate}
\title{mutation_rate}
\format{
matrix size 4x5
}
\source{
https://www.agilent.com/cs/library/usermanuals/public/200550.pdf
}
\usage{
mutation_rate
}
\description{
Mutational rate matrix for error-prone PCR, obtained from GeneMorph II Random Mutagenesis Kit.
}
\keyword{datasets}
