% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_prior_errors.r
\name{p_prior_errors}
\alias{p_prior_errors}
\title{Computes prior probability of errors}
\usage{
p_prior_errors(counts_pnq, output_file = NULL, save = FALSE)
}
\arguments{
\item{counts_pnq}{Data frame with columns position nucleoide quality counts, as returned by parse_countspnq}

\item{output_file}{File name for output, if save=TRUE.}

\item{save}{Logical. Should data be saved in a output_file?}
}
\value{
Data frame with columns position nucleotide prior.error.
}
\description{
This is an auxiliary function in single package. It takes a data frame with counts by position, nucleotide and Qscore and it summarises it into proportion of nucleotide counts by position.
}
\examples{
refseq_fasta <- system.file("extdata", "ref_seq.fasta", package = "single")
train_reads_example <- system.file("extdata", "train_seqs_500.sorted.bam",
                                   package = "single")
counts_pnq <- pileup_by_QUAL(train_reads_example,pos_start=1,pos_end=10)
p_prior_errors <- p_prior_errors(counts_pnq=counts_pnq)
head(p_prior_errors)
}
