% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_prior_mutations.r
\name{p_prior_mutations}
\alias{p_prior_mutations}
\title{Computes prior probability of mutations}
\usage{
p_prior_mutations(
  rates.matrix,
  mean.n.mut,
  ref_seq,
  save = FALSE,
  output_file = "tablePriorMutations.txt"
)
}
\arguments{
\item{rates.matrix}{Mutation rate matrix: 4x5 matrix, each row/col representing a nucleotide (col adds deletion), and the values is the mutational rate from row to col.}

\item{mean.n.mut}{Mean number of mutations expected (one number).}

\item{ref_seq}{Reference sequence: vector of characters, as returned by load_ref_seq}

\item{save}{Logical. Should data be saved in a output_file?}

\item{output_file}{File name for output, if save=TRUE.}
}
\value{
Data frame with columns wt.base (wild type nucleotide), nucleotide (mutated nucleotide), p_mutation (probaility of mutation)
}
\description{
This is an auxiliary function in single package. It computes the prior probability of mutation in a gene library.
}
\examples{
pos_start <- 1
pos_end <- 10
refseq_fasta <- system.file("extdata", "ref_seq.fasta", package = "single")
train_reads_example <- system.file("extdata", "train_seqs_500.sorted.bam",
                                   package = "single")
counts_pnq <- pileup_by_QUAL(bam_file=train_reads_example,
    pos_start=pos_start,pos_end=pos_end)
p_prior_mutations <- p_prior_mutations(rates.matrix = mutation_rate,
    mean.n.mut = 5,ref_seq = ref_seq,save = F)
head(p_prior_mutations)
}
