% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_countspnq.r
\name{parse_countspnq}
\alias{parse_countspnq}
\title{Parse nucleotide' counts file}
\usage{
parse_countspnq(
  input_file,
  output_file,
  pos_start = NULL,
  pos_end = NULL,
  save = FALSE
)
}
\arguments{
\item{input_file}{File containing the counts per position returned by samtools mpileup}

\item{output_file}{Name of file to save parsed data.}

\item{pos_start}{Numeric. Position to start analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{pos_end}{Numeric. Position to stop analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{save}{Logical. Should data be saved in a output_file?}
}
\value{
data frame with columns position nucleotide quality counts
}
\description{
This is an auxiliary function in single package. Reshapes file returned by samtools mpileup into a data.frame.
}
\examples{
file_ref_pnq = system.file("extdata", "REF_READS_PNQ.txt", package = "single")
parse_countspnq(input_file=file_ref_pnq,output_file=NA, pos_end=10)
}
