% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nucleotides_per_qscore.r
\name{parse_nucleotides_per_qscore}
\alias{parse_nucleotides_per_qscore}
\title{Parse nucleotide' counts file}
\usage{
parse_nucleotides_per_qscore(
  input_file,
  output_file,
  pos_start = NULL,
  pos_end = NULL,
  save = FALSE
)
}
\arguments{
\item{input_file}{File containing the counts per position returned by samtools mpileup}

\item{output_file}{Name of file to save parsed data.}

\item{pos_start}{Numeric. Position to start analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{pos_end}{Numeric. Position to stop analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{save}{Logical. Should data be saved in a output_file?}
}
\value{
data frame with columns position nucleotide quality counts
}
\description{
This is an auxiliary function in single package. Reshapes file returned by samtools mpileup into a data.frame.
}
\examples{
input_file = system.file("extdata", "example_train_countsPNQ.txt", package = "single")
parse_nucleotides_per_qscore(input_file=input_file,output_file=NA, pos_start=1,pos_end=10)
}
