% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_errors.r
\name{prior_errors}
\alias{prior_errors}
\title{Computes prior probability of errors}
\usage{
prior_errors(counts_pnq, output_file = NULL, ref_seq, save = FALSE)
}
\arguments{
\item{counts_pnq}{Data frame with columns position nucleoide quality counts, as returned by parse_countspnq}

\item{output_file}{File name for output, if save=TRUE.}

\item{ref_seq}{Reference sequence: vector of characters, as returned by load_ref_seq}

\item{save}{Logical. Should data be saved in a output_file?}
}
\value{
Data frame with columns position nucleotide prior.error.
}
\description{
This is an auxiliary function in single package. It takes a data frame with counts by position, nucleotide and Qscore and it summarises it into proportion of nucleotide counts by position.
}
\examples{
refseq_file = system.file("extdata", "ref_seq.fasta", package = "single")
ref_seq = load_ref_seq(refseq_file)
file_ref_pnq = system.file("extdata", "REF_READS_PNQ.txt", package = "single")
counts_pnq = parse_countspnq(input_file=file_ref_pnq,output_file=NA)
prior_errors(counts_pnq,ref_seq=ref_seq)
}
