% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_consensus_byBarcode.r
\name{single_consensus_byBarcode}
\alias{single_consensus_byBarcode}
\title{Compute SINGLE consensus}
\usage{
single_consensus_byBarcode(
  barcodes_table_file,
  reference_sequence_fastafile,
  single_corrected_files,
  file_out_consensus,
  column_id = NULL,
  column_barcode = NULL,
  header = TRUE,
  dec = ".",
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{barcodes_table_file}{File containing the names of the reads and the barcode associated (or any grouping tag).}

\item{reference_sequence_fastafile}{Fasta file containing reference sequence}

\item{single_corrected_files}{Files containing the sequences corrected by SINGLE, i.e. saved by single_evaluate_model.}

\item{file_out_consensus}{String. File to store results.}

\item{column_id, column_barcode}{Numeric. Columns where the sequences id and the barcode (or grouping tag) are, in the barcodes_table_file}

\item{header, dec, sep}{Arguments for read.table(barcodes_table_file)}

\item{verbose}{Logical.}
}
\value{
This function returns 0. Results are saved in file_out_consensus
}
\description{
Main function to compute consensus after correcting reads by a SINGLE model.
}
\examples{
barcodes_table_file = system.file("extdata", "example_barcodes_table.txt", package = "single")
ref_seq_file = system.file("extdata", "ref_seq.fasta", package = "single")
single_corrected_files = system.file("extdata", "example_corrected.txt", package="single")
file_out = tempfile("example_consensus.txt")
single_consensus_byBarcode(barcodes_table_file,ref_seq_file,single_corrected_files,file_out)
}
