% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_consensus.r
\name{weighted_consensus}
\alias{weighted_consensus}
\title{Compute consensus sequence}
\usage{
weighted_consensus(df, cutoff_prob = 0.2)
}
\arguments{
\item{df}{data.frame with the columns: nucleotide, probability, position}

\item{cutoff_prob}{Numeric. Nucleotides with probability below this number will be removed from consensus computation.}
}
\value{
Character vector, consensus sequence
}
\description{
This is an auxiliary function in single package. It computes consensus from a data.frame as the one returned by single_evaluate()
}
\examples{
fastq_seqs_example <- system.file("extdata", "test_sequences.fastq",package = "single")
seqs_example <- Biostrings::readQualityScaledDNAStringSet(fastq_seqs_example)
# Using single weights
data_barcode = data.frame(
 nucleotide = unlist(sapply(as.character(seqs_example),strsplit, split="")),
 p_SINGLe=unlist(1-as(Biostrings::quality(seqs_example),"NumericList")),
 pos=rep(1:Biostrings::width(seqs_example[1]),length(seqs_example)))
weighted_consensus(df = data_barcode, cutoff_prob = 0.9)
# Replacing weights by ones
data_barcode = data.frame(
 nucleotide = unlist(sapply(as.character(seqs_example),strsplit, split="")),
 p_SINGLe=1,pos=rep(1,sum(Biostrings::width(seqs_example))))
weighted_consensus(df = data_barcode, cutoff_prob = 0)
}
