% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup_by_QUAL.R
\name{pileup_by_QUAL}
\alias{pileup_by_QUAL}
\title{Pileup by QUAL}
\usage{
pileup_by_QUAL(
  bam_file,
  QUAL_values = seq(93, 0),
  pos_start = NA,
  pos_end = NA
)
}
\arguments{
\item{bam_file}{Bam file to pile up}

\item{QUAL_values}{Numeric vector. QUAL values to analyze in the data.}

\item{pos_start}{Numeric. Position to start analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{pos_end}{Numeric. Position to stop analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}
}
\value{
data.frame with columns strand,pos,nucleotide,QUAL,countss
}
\description{
To explain
}
\examples{
refseq_fasta <- system.file("extdata", "ref_seq.fasta", package = "single")
train_reads_example <- system.file("extdata", "train_seqs_500.sorted.bam",
                                   package = "single")
counts_pnq <- pileup_by_QUAL(bam_file=train_reads_example,
    pos_start=1,pos_end=10)
head(counts_pnq)
}
