% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_txt_file.r
\name{save_txt_file}
\alias{save_txt_file}
\title{Save evaluated SINGLE model}
\usage{
save_txt_file(
  input_sam_file,
  single_fits,
  output_txt_file,
  ref_seq,
  pos_start = NULL,
  pos_end = NULL,
  comment.char = "@",
  gaps_weights,
  remove.first.reverse = FALSE
)
}
\arguments{
\item{input_sam_file}{File containing the counts per position returned by samtools mpileup}

\item{single_fits}{Results of the SINGLE model as returned by single_train(). It can be either the output data.frame or the saved file.}

\item{output_txt_file}{String. Prefix for output files}

\item{ref_seq}{Reference sequence: vector of characters, as returned by load_ref_seq}

\item{pos_start}{Numeric. Position to start analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{pos_end}{Numeric. Position to stop analyzing, counting starts from 1 and it refers to reference used for minimap2 alignment.}

\item{comment.char}{Character used to indicate comments in the input_sam_file}

\item{gaps_weights}{One of "minimum","none","mean". How to assign qscores to deletions.}

\item{remove.first.reverse}{Logical. Do you want to remove first position in the reverse reads?}
}
\value{
Writes txt file
}
\description{
This is an auxiliary function in single package. It saves the results of the evaluation in a txt file.
}
