% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_consensus_byBarcode.r
\name{single_consensus_byBarcode}
\alias{single_consensus_byBarcode}
\title{Compute SINGLE consensus}
\usage{
single_consensus_byBarcode(
  barcodes_file,
  single_corrected_seqs,
  column_SeqID = NULL,
  column_BCid = NULL,
  header = TRUE,
  dec = ".",
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{barcodes_file}{File containing the names of the reads and the barcode associated (or any grouping tag).}

\item{single_corrected_seqs}{Files containing the sequences corrected by SINGLE, i.e. saved by single_evaluate}

\item{column_SeqID, column_BCid}{Numeric. Columns where the sequences id and the barcode (or grouping tag) are, in the barcodes_file}

\item{header, dec, sep}{Arguments for read.table(barcodes_file)}

\item{verbose}{Logical.}
}
\value{
DNAStringSet with consensus sequences
}
\description{
Main function to compute consensus after correcting reads by a SINGLE model.
}
\examples{
pos_start=1
pos_end = 10
refseq_fasta = system.file("extdata", "ref_seq.fasta", package = "single")
ref_seq <- subseq(Biostrings::readDNAStringSet(refseq_fasta), pos_start,pos_end)
train_reads_example <- system.file("extdata", "train_seqs_500.sorted.bam",
                                   package = "single")
train <- single_train(bamfile=train_reads_example,
                   refseq_fasta=refseq_fasta,
                   rates.matrix=mutation_rate,
                   mean.n.mutations=5.4,
                   pos_start=pos_start,
                   pos_end=pos_end,
                   save_final= FALSE)
test_reads_example = system.file("extdata", "test_sequences.sorted.bam",
   package = "single")
corrected_reads <- single_evaluate(bamfile = test_reads_example,
                 single_fits = train,
                 ref_seq = ref_seq,
                 pos_start=pos_start,
                 pos_end=pos_end,
                 gaps_weights = "minimum")
barcodes_table_example = system.file("extdata", "Barcodes_table.txt",
   package = "single")
consensus <- single_consensus_byBarcode(
                 barcodes_file = barcodes_table_example,
                 single_corrected_seqs = corrected_reads,
                 verbose = FALSE)
}
