% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_consensus_byBarcode.r
\name{single_consensus_byBarcode}
\alias{single_consensus_byBarcode}
\title{Compute SINGLE consensus}
\usage{
single_consensus_byBarcode(
  barcodes_file,
  refseq_fasta,
  single_corrected_seqs,
  column_id = NULL,
  column_barcode = NULL,
  header = TRUE,
  dec = ".",
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{barcodes_file}{File containing the names of the reads and the barcode associated (or any grouping tag).}

\item{refseq_fasta}{Fasta file containing reference sequence}

\item{single_corrected_seqs}{Files containing the sequences corrected by SINGLE, i.e. saved by single_evaluate}

\item{column_id, column_barcode}{Numeric. Columns where the sequences id and the barcode (or grouping tag) are, in the barcodes_file}

\item{header, dec, sep}{Arguments for read.table(barcodes_file)}

\item{verbose}{Logical.}
}
\value{
DNAStringSet with consensus sequences
}
\description{
Main function to compute consensus after correcting reads by a SINGLE model.
}
\examples{
pos_start=1
pos_end = 100
barcodes_file = system.file("extdata", "BC_TABLE.txt", package = "single")
refseq_file = system.file("extdata", "ref_seq.fasta", package = "single")
reads_single = system.file("extdata", "single_reads_ex_corrected.txt", package = "single")
single_consensus_byBarcode(barcodes_file,refseq_file,reads_single,
                          verbose = FALSE)
}
