% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_consensus_byBarcode.r
\name{single_consensus_byBarcode}
\alias{single_consensus_byBarcode}
\title{Compute SINGLE consensus}
\usage{
single_consensus_byBarcode(
  barcodes_table,
  sequences,
  readID_col = 1,
  bcID_col = 2,
  header = TRUE,
  dec = ".",
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{barcodes_table}{data.frame or file name containing the names of the reads and the barcode associated (or any grouping tag).}

\item{sequences}{QualityScaledDNAStringSet or fastq file name. Contains sequences from which compute weighted consensus.}

\item{readID_col, bcID_col}{Numeric. Columns where the reads id and the barcode (or grouping tag) are, in the barcodes_table}

\item{header, dec, sep}{Arguments for read.table(barcodes_table)}

\item{verbose}{Logical.}
}
\value{
DNAStringSet with consensus sequences
}
\description{
Main function to compute consensus after correcting reads by a SINGLE model.
}
\examples{
pos_start=1
pos_end = 100
barcodes_file =  system.file("extdata", "Barcodes_table.txt",package = "single")
reads_single = system.file("extdata", "corrected_seqs.fastq", package = "single")
single_consensus_byBarcode(barcodes_file,reads_single, verbose = FALSE)
}
