% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_consensus.r
\name{weighted_consensus}
\alias{weighted_consensus}
\title{Compute consensus sequence}
\usage{
weighted_consensus(df, cutoff_prob = 0.2)
}
\arguments{
\item{df}{data.frame with the columns: nucleotide, probability, position}

\item{cutoff_prob}{Numeric. Nucleotides with probability below this number will be removed from consensus computation.}
}
\value{
Character vector, consensus sequence
}
\description{
This is an auxiliary function in single package. It computes consensus from a data.frame as the one returned by single_evaluate()
}
\examples{
data_barcode = data.frame(
 nucleotide = unlist(sapply(as.character(corrected_reads),strsplit, split="")),
 p_SINGLe=unlist(1-as(quality(corrected_reads),"NumericList")),
 pos=rep(1:width(corrected_reads[1]),length(corrected_reads)))
weighted_consensus(df = data_barcode, cutoff_prob = 0.9)
weighted_consensus(df = data_barcode, cutoff_prob = 0.999)
}
